Additional/requirements.txt
numpy
matplotlib
scipy
torch
librosa
tqdm
transformers
torchaudio
confidence_intervals
hydra-core
lightning
pandas
wandb